+++ ParticleAnnotation/utils/model/active_learning_model.py
@@ -82,7 +82,8 @@ def predict_3d_with_AL(
             ] = logits
 
     # Extract peaks
-    max_filter = maximum_filter(full_logits, size=maximum_filter_size)
+    max_filter = maximum_filter(full_logits.astype(np.float32), 
+                                size=maximum_filter_size)
     peaks = full_logits - max_filter
     peaks = np.where(peaks == 0)
     peaks = np.stack(peaks, axis=-1)
